delete playerAI, updated Chessboard, GameApplication call AI, enums value of pieces, tests

+++ Chess/src/main/java/com/chess/core/enums/TypePiece.java
@@ -8,7 +8,7 @@ public enum TypePiece {
 	
 	private double value;
 	
-	TypePiece(int value){
+	TypePiece(double value){
 		this.value = value;
 	}
